submit trun cart

+++ client-react/src/pages/Customized/components/RightAside/RightAside.js
@@ -1,6 +1,7 @@
 import React from "react";
 import CartListContent from "./CartListContent";
 import Swal from "sweetalert2";
+import {withRouter} from 'react-router-dom'
 
 function RightAside(props) {
   const {
@@ -14,22 +15,8 @@ function RightAside(props) {
     setTotalCart,
     setCartUpdate,
   } = props;
-    
-  const customizedProduct = {
-    productId: 99,
-    productName: "客製化水果盒",
-    count:1,
-    amount: "",
-    price: totalPrice,
-    imageUrl:"http://localhost:3000/images/CustomizedPhotos/customized.jpeg"
-  };
-  const amount = [];
-  for (let i = 0; i < cartData.length; i++) {
-    amount.push(`${cartData[i].fruitname}*${counts[i]}`);
-  }
-  customizedProduct.amount = amount.join(",");
 
-  function sweetAlert(){
+  function successAdd() {
     Swal.fire({
       title: `${customizedProduct.productName}加入成功`,
       text: "點擊右上角查看",
@@ -38,28 +25,52 @@ function RightAside(props) {
       confirmButtonText: "關閉",
     });
   }
+  function warning() {
+    Swal.fire({
+      title: "您還沒選擇商品",
+      confirmButtonText: "關閉",
+    });
+  }
 
+  const customizedProduct = {
+    productId: 99,
+    productName: "客製化水果盒",
+    count: 1,
+    amount: "",
+    price: totalPrice,
+    imageUrl: "http://localhost:3000/images/CustomizedPhotos/customized.jpeg",
+  };
+  const amount = [];
+  for (let i = 0; i < cartData.length; i++) {
+    amount.push(`${cartData[i].fruitname}*${counts[i]}`);
+  }
+  customizedProduct.amount = amount.join(",");
 
+  const updateCartToLocalStorage = () => {
+    if (cartData.length === 0) {
+      warning();
+    } else {
+      const currentCart = JSON.parse(localStorage.getItem("cart")) || [];
+      currentCart.push(customizedProduct);
+      localStorage.setItem("cart", JSON.stringify(currentCart));
 
-const updateCartToLocalStorage = () => {
-  const currentCart = JSON.parse(localStorage.getItem('cart')) || []
-  currentCart.push(customizedProduct)
-  localStorage.setItem('cart', JSON.stringify(currentCart))
+      setTotalCart(function (prevData) {
+        const newTotalCart = [...prevData];
+        newTotalCart.push(customizedProduct);
+        return newTotalCart;
+      });
+      setCartUpdate(true);
+      setCartData([]);
+      setCounts([]);
+      successAdd();
+    }
+  };
+  function addCartAndTurnCartPage(){
+    updateCartToLocalStorage()
+    props.history.push('/cart')
+  }
 
-  setTotalCart(function (prevData) {
-    const newTotalCart = [...prevData];
-    newTotalCart.push(customizedProduct)
-    return newTotalCart;
-  });
-  setCartUpdate(true)
-  setCartData([])
-  setCounts([])
-  sweetAlert()
-  
-}
-
-
-return (
+  return (
     <>
       <aside className="col-md-4 d-none d-lg-block position-relative">
         <h2 className="text-center customerCartList">客製化列表</h2>
@@ -91,6 +102,7 @@ return (
                 wight={wight}
                 nutrients={nutrients}
                 index={index}
+                cartData={cartData}
                 setCartData={setCartData}
                 cartItem={cartData[index]}
                 setCounts={setCounts}
@@ -121,10 +133,10 @@ return (
         </div>
         <div className="d-flex justify-content-around">
           <button onClick={updateCartToLocalStorage} className="buy-btn">
-            <i  className="fas fa-shopping-cart"></i>
+            <i className="fas fa-shopping-cart"></i>
             加入購物車
           </button>
-          <button className="sub">立即結帳</button>
+          <button onClick={addCartAndTurnCartPage} className="sub">立即結帳</button>
         </div>
         <button className="btn removeFruit">Remove All</button>
         <div className="fruitbox"></div>
@@ -132,4 +144,4 @@ return (
     </>
   );
 }
-export default RightAside;
+export default withRouter(RightAside);
